+++ src/changelog.ts
@@ -1,6 +1,10 @@
 export const latestChangelog = {
     version: '0.4.0',
     date: '2025-11-13',
-    features: ['New interface layout', 'Overlay "discard" on cards in hand'],
+    features: [
+        'New interface layout',
+        'User-defined keyboard shortcuts',
+        'Overlay "discard" on cards in hand',
+    ],
     bugfixes: ['Fix markers display', 'Fix bot stealth & blocking'],
 }